Refactoring

+++ engine/include/bot.h
@@ -8,12 +8,12 @@ class Bot {
 private:
     int evaluate(ChessLogic &);
     int negamax(ChessLogic &, int, int, int, int);
-    u64 search_move(ChessLogic &, int, int);
-    std::vector<u64> order_moves(ChessLogic &, std::vector<u64>);
+    u16 search_move(ChessLogic &, int, int);
+    std::vector<u16> order_moves(ChessLogic &, std::vector<u16>);
 
 public:
     Bot();
-    u64 choose_move(ChessLogic, std::vector<u64> &);
+    u16 choose_move(ChessLogic, std::vector<u16> &);
 };
 
 #endif // !BOT_H
